requirements.txt
#pip install -r ./requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
boto==2.49.0
boto3==1.18.17
botocore==1.21.17
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
dj-database-url==0.5.0
Django==2.2.1
django-braces==1.14.0
django-cors-headers==3.7.0
django-crontab==0.7.1
django-oauth-toolkit==1.5.0
django-rest-framework-social-oauth2==1.1.0
django-storages==1.11.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
gunicorn==20.1.0
h11==0.13.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jwcrypto==0.9.1
lxml==4.8.0
MarkupSafe==2.0.1
oauthlib==3.1.1
outcome==1.1.0
Pillow==8.3.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.1
trio==0.20.0
trio-websocket==0.9.2
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
wrapt==1.12.1
wsproto==1.1.0
